+++ test/utest-online/testlinkapi_generic_online_test.py
@@ -441,10 +441,17 @@ class TestLinkAPIGenericOnlineTestCase(unittest.TestCase):
                             {'cf_tc_ex_string' : 'a custom exec value', 
                              'cf_tc_ex_numeric' : 111} )
 
+    def test_getTestSuite_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, 'NO.*TProjectPrefix'):
+            self.client.getTestSuite('suite 40000712', 'TProjectPrefix')
+            
+
     def test_updateTestSuite_unknownID(self):
         with self.assertRaisesRegex(TLResponseError, '7000.*40000711'):
-            self.client.updateTestSuite('suite 40000712', testprojectid=40000711, 
-                                        details='detail 40000713')
+            self.client.updateTestSuite(40000712, testprojectid=40000711, 
+                                    testsuitename = 'suite 40000712 updated',
+                                    details = 'detail 40000713 updated',
+                                    order =1)